+++ package/main_pygame.py
@@ -4,14 +4,19 @@ import pygame
 from pygame.locals import *
 
 import live2d.v3 as live2d
+import live2d.log as log
 import resouces
-from wav_handler import WavHandler
 
 # import live2d.v2 as live2d
 
 live2d.setLogEnable(True)
 
-model: live2d.LAppModel
+
+def on_start_motion_callback(group: str, no: int):
+    log.Info("start motion: [%s_%d]" % (group, no))
+
+def on_finish_motion_callback():
+    log.Info("motion finished")
 
 
 def draw():
@@ -20,13 +25,10 @@ def draw():
 
 
 def main():
-    global model
-
     pygame.init()
     pygame.mixer.init()
     live2d.init()
 
-    wavHandler = WavHandler()
 
     display = (700, 500)
     pygame.display.set_mode(display, DOUBLEBUF | OPENGL)
@@ -37,7 +39,6 @@ def main():
 
     model = live2d.LAppModel()
 
-    model = live2d.LAppModel()
     if live2d.LIVE2D_VERSION == 3:
         model.LoadModelJson(os.path.join(resouces.RESOURCES_DIRECTORY, "v3/Haru/Haru.model3.json"))
     else:
@@ -51,23 +52,10 @@ def main():
     dy: float = 0.0
     scale: float = 1.0
 
-    cnt = 0
-
-    # 关闭内置口型同步
-    model.SetLipSyncEnable(False)
-
     # 关闭自动眨眼
     # model.SetAutoBlinkEnable(False)
     # 关闭自动呼吸
     # model.SetAutoBreathEnable(False)
-
-    def start_callback(group, no):
-        print("start lipsync")
-        pygame.mixer.music.load("audio1.wav")
-        pygame.mixer.music.play()
-        wavHandler.Start("audio1.wav")
-
-    lipSyncN = 3
 
     while True:
         for event in pygame.event.get():
@@ -76,10 +64,9 @@ def main():
                 break
             if event.type == pygame.MOUSEBUTTONDOWN:
                 x, y = pygame.mouse.get_pos()
-                model.Touch(x, y)
+                model.Touch(x, y, on_start_motion_callback, on_finish_motion_callback)
 
             if event.type == pygame.KEYDOWN:
-                print(event.key)
                 if event.key == pygame.K_LEFT:
                     dx -= 0.1
                 elif event.key == pygame.K_RIGHT:
@@ -100,17 +87,11 @@ def main():
             if event.type == pygame.MOUSEMOTION:
                 model.Drag(*pygame.mouse.get_pos())
 
+
         if not running:
             break
 
         model.CalcParameters()
-        if cnt < 1:
-            cnt += 1
-            model.StartMotion("Speak", 0, live2d.MotionPriority.FORCE.value, start_callback)
-
-        if wavHandler.Update():  # 当前音频仍在播放
-            model.SetParameterValue("ParamMouthOpenY", wavHandler.GetRms() * lipSyncN, 1)
-
         model.SetOffset(dx, dy)
         model.SetScale(scale)
         live2d.clearBuffer()
