+++ package/resouces.py
@@ -1,4 +1,4 @@
 import os
 
-__current_directory = os.path.split(__file__)[0]
-RESOURCES_DIRECTORY = os.path.join(__current_directory, "../Resources")
+CURRENT_DIRECTORY = os.path.split(__file__)[0]
+RESOURCES_DIRECTORY = os.path.join(CURRENT_DIRECTORY, "../Resources")
